GCC2 build fixes on radeon_hd driver
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@41858 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/accelerants/radeon_hd/mode.cpp
@@ -189,7 +189,7 @@ CardFBSet(display_mode *mode)
 	uint64_t fbAddress = gInfo->shared_info->frame_buffer_phys;
 
 	// Tell GPU which frame buffer address to draw from
-	if (gInfo->shared_info->device_chipset >= (RADEON_R700 & 0x70)) {
+	if (gInfo->shared_info->device_chipset >= (uint16)RADEON_R700 & 0x70) {
 		write32(gRegister->grphPrimarySurfaceAddrHigh,
 			(fbAddress >> 32) & 0xf);
 		write32(gRegister->grphSecondarySurfaceAddrHigh,
